+++ src/main/kotlin/mqtt/Utils.kt
@@ -3,28 +3,28 @@ package mqtt
 import java.io.InputStream
 import java.io.OutputStream
 
-fun OutputStream.encodeVariableByteInteger(value: Int) {
+fun OutputStream.encodeVariableByteInteger(value: UInt) {
     var x = value
     do {
-        var encodedByte = x.rem(128)
-        x /= 128
-        if (x > 0) {
-            encodedByte = encodedByte or 128
+        var encodedByte = x.rem(128u)
+        x /= 128u
+        if (x > 0u) {
+            encodedByte = encodedByte or 128u
         }
-        write(encodedByte)
-    } while (x > 0)
+        write(encodedByte.toInt())
+    } while (x > 0u)
 }
 
-fun InputStream.decodeVariableByteInteger(): Int {
-    var multiplier = 1
-    var value = 0
+fun InputStream.decodeVariableByteInteger(): UInt {
+    var multiplier = 1u
+    var value = 0u
     do {
-        val encodedByte = read()
-        value += (encodedByte and 127) * multiplier
-        if (multiplier > 128 * 128 * 128) {
+        val encodedByte = read().toUInt()
+        value += (encodedByte and 127u) * multiplier
+        if (multiplier > 128u * 128u * 128u) {
             throw Exception("Malformed Variable Byte Integer")
         }
-        multiplier *= 128
-    } while ((encodedByte and 128) != 0)
+        multiplier *= 128u
+    } while ((encodedByte and 128u) != 0u)
     return value
 }
